requirements.txt
streamlit>=1.35
pandas>=2.2
pillow>=10.3
